+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Show "Badges" and "Settings" in top navigation (#234)
 - Upgrade to Django 2.2
 - Can configure the email integration to only report the "down" events (#231)
+- Add "Test!" function in the Integrations page (#207)
 
 
 ## 1.6.0 - 2019-04-01